(PC-25868) refactor(offer): camera tag display
